pyproject.toml
[project]
name = "fast_auto_scrna"
version = "2.0.0.dev0"
description = "End-to-end single-cell RNA-seq atlas pipeline with Rust-accelerated kernels"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Phoenix12580" }]

dependencies = [
    "anndata>=0.10",
    "scanpy>=1.11",
    "numpy>=1.26",
    "scipy>=1.11",
    "pandas>=2.0",
    "scikit-learn>=1.4",
    "scikit-misc>=0.5",      # loess for seurat_v3 HVG + ROGUE entropy fit
    "statsmodels>=0.14",     # multitest FDR for ROGUE
    "h5py>=3.10",
    "python-igraph>=0.11",
    "leidenalg>=0.10",
    "umap-learn>=0.5",
    "hnswlib>=0.8",
    "rdata>=0.9",
    "matplotlib>=3.8",
    "scib-metrics>=0.5",     # JAX-jitted ASW (label/batch/isolated) — paper-faithful
    "scvi-tools>=1.1",       # SCVI VAE batch-correction route (PyTorch backend)
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-xdist>=3", "ruff>=0.5"]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fast_auto_scrna"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
